# tmi-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TMI (Threat Modeling Improved) API",
    "description": "A RESTful API for collaborative threat modeling with full X6 graph library compatibility. This API provides schemas that align with AntV X6 cell object models for seamless integration with modern diagramming libraries. Supports OAuth 2.0 authentication with client callback integration for seamless single-page application authentication flows.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://api.tmi.example.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated requests"
      },
      "websocketAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "JWT token for WebSocket authentication. WebSocket connections must include the JWT token as a query parameter since browsers do not support custom headers for WebSocket connections."
      }
    },
    "schemas": {
...
      "Metadata": {
        "type": "object",
        "description": "A key-value pair for extensible metadata",
        "properties": {
          "key": {
            "type": "string",
            "description": "Metadata key",
            "minLength": 1,
            "maxLength": 256,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "x-oapi-codegen-extra-tags": {
              "binding": "required"
            }
          },
          "value": {
            "type": "string",
            "description": "Metadata value",
            "minLength": 1,
            "maxLength": 1024,
            "pattern": "^[^<>]*$",
            "x-oapi-codegen-extra-tags": {
              "binding": "required"
            }
          }
        },
        "required": ["key", "value"]
      },
      "Error": {
        "type": "object",
        "description": "Standard error response format",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code"
          },
          "error_description": {
            "type": "string",
            "description": "Human-readable error description"
          },
          "error_uri": {
            "type": "string",
            "format": "uri",
            "description": "URI to documentation about the error"
          }
        },
        "required": ["error", "error_description"],
        "example": {
          "error": "invalid_request",
          "error_description": "The request is missing a required parameter",
          "error_uri": "https://docs.example.com/auth/errors/invalid_request"
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "authorizationError": {
                "summary": "Insufficient permissions for sub-resource access",
                "value": {
                  "error": "access_denied",
                  "error_description": "User does not have sufficient permissions to access this sub-resource",
                  "details": {
                    "required_role": "writer",
                    "user_role": "reader",
                    "threat_model_id": "550e8400-e29b-41d4-a716-446655440000",
                    "resource_type": "threat"
                  }
                }
              },
              "notFoundError": {
                "summary": "Sub-resource not found",
                "value": {
                  "error": "not_found",
                  "error_description": "The requested sub-resource does not exist or does not belong to the specified threat model",
                  "details": {
                    "resource_id": "550e8400-e29b-41d4-a716-446655440999",
                    "resource_type": "document",
                    "threat_model_id": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              },
              "validationError": {
                "summary": "Invalid input data",
                "value": {
                  "error": "invalid_input",
                  "error_description": "The provided data does not meet validation requirements",
                  "details": {
                    "field_errors": [
                      { "field": "name", "message": "Name is required and cannot be empty" },
                      { "field": "url", "message": "URL must be a valid HTTP/HTTPS URL" }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/threat_models/{threat_model_id}/sources/{source_id}/metadata": {
      "get": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Get source metadata",
        "description": "Returns all metadata key-value pairs for the specified source reference",
        "operationId": "getSourceMetadata",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the source reference (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "Source metadata list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Metadata"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Create source metadata",
        "description": "Creates new metadata entry for the specified source reference",
        "operationId": "createSourceMetadata",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the source reference (UUID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Metadata"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Metadata entry created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Metadata"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/threat_models/{threat_model_id}/sources/{source_id}/metadata/{key}": {
      "get": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Get source metadata by key",
        "description": "Returns a specific metadata entry by key for the specified source reference",
        "operationId": "getSourceMetadataByKey",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the source reference (UUID)"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Metadata key to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Metadata"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "put": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Update source metadata by key",
        "description": "Updates a specific metadata entry by key for the specified source reference",
        "operationId": "updateSourceMetadataByKey",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the source reference (UUID)"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Metadata key to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "string",
                    "description": "New value for the metadata entry"
                  }
                },
                "required": ["value"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadata entry updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Metadata"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Delete source metadata by key",
        "description": "Deletes a specific metadata entry by key for the specified source reference",
        "operationId": "deleteSourceMetadataByKey",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the source reference (UUID)"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Metadata key to delete"
          }
        ],
        "responses": {
          "204": { "description": "Metadata entry deleted successfully" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/threat_models/{threat_model_id}/sources/{source_id}/metadata/bulk": {
      "post": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Bulk create source metadata",
        "description": "Creates multiple metadata entries in a single operation for the specified source reference",
        "operationId": "bulkCreateSourceMetadata",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the source reference (UUID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Metadata"
                },
                "maxItems": 20
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Metadata entries created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Metadata"
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  }
}